requirements.txt
# ── Core RAG (0.10.x ecosystem — matches codebase) ─────────────────
llama-index-core==0.10.68
llama-index-embeddings-huggingface==0.1.4
llama-index-storage-docstore-redis==0.1.2
llama-index-vector-stores-redis==0.1.2
llama-index-llms-openai==0.1.31

# ── Multi-LLM (RAGSynapse v2 additions — 0.10.x compatible) ────────
llama-index-llms-anthropic==0.1.15
llama-index-llms-ollama==0.1.5
llama-index-agent-openai==0.2.9
openai==1.40.0
langchain-ollama==0.1.3

# ── Embeddings — compatible trio ────────────────────────────────────
sentence-transformers==3.0.1
transformers==4.41.2
accelerate==0.30.0
huggingface-hub==0.23.4
# torch is installed separately in Dockerfile (CPU wheel)

# ── UI ───────────────────────────────────────────────────────────────
streamlit==1.27.2

# ── Document Processing ──────────────────────────────────────────────
PyPDF2==3.0.1
PyMuPDF==1.23.8
pdf2image==1.17.0
pytesseract==0.3.10
python-docx==1.1.0
pillow==10.2.0
fpdf==1.7.2

# ── Config ───────────────────────────────────────────────────────────
python-dotenv==1.0.0
toml==0.10.2

# ── Observability (Week 2/3) ─────────────────────────────────────────
mlflow==2.9.2
ragas==0.1.9
langchain-community==0.2.16

# ── HTTP ─────────────────────────────────────────────────────────────
httpx==0.27.2